Added Cloudinary Overlay Lab
This is a new tool for helpers to upload raw video to Cloudinary and export an eagerly transformed video.
While the source code is readily available here, including the signing mechanism which is documented on Cloudinary's website, the API secret is required to finish uploading to the server. Interested developers can simply use their own credentials and eager transformations to try this lab.